# WinCCOAGraphQLRestApisExample/javascript/winccoa-graphql-server-main/restapi/openapi-full.yaml
openapi: 3.0.0
info:
  title: WinCC OA REST API
  version: 1.0.0
  description: |
    REST API for WinCC OA GraphQL Server - provides HTTP endpoints for all WinCC OA operations including:
    - Data point management (CRUD, value operations)
    - Data point type management
    - Tag queries with metadata
    - Alert handling
    - CNS (Central Navigation Service) operations
    - System information and redundancy
    - OPC UA configuration
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /restapi/v1/auth/login
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Missing required parameter
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ElementType:
      type: string
      enum:
        - BOOL
        - UINT8
        - INT32
        - INT64
        - FLOAT
        - DOUBLE
        - BIT
        - BIT32
        - BIT64
        - STRING
        - TIME
        - DPID
        - LANGSTRING
        - BLOB
        - MIXED
        - DYN_BOOL
        - DYN_UINT8
        - DYN_INT32
        - DYN_INT64
        - DYN_FLOAT
        - DYN_DOUBLE
        - DYN_BIT
        - DYN_BIT32
        - DYN_BIT64
        - DYN_STRING
        - DYN_TIME
        - DYN_DPID
        - DYN_LANGSTRING
        - DYN_BLOB
...
    DpTypeNode:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          example: value
        type:
          $ref: '#/components/schemas/ElementType'
        refName:
          type: string
          example: RefType
        children:
          type: array
          items:
            $ref: '#/components/schemas/DpTypeNode'
        newName:
          type: string
          example: newValue
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Missing or invalid authentication token
    Forbidden:
      description: Insufficient permissions (admin role required)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: Admin role required
paths:
...
  /restapi/v1/datapoints:
    get:
      tags: [Data Points]
      summary: List data points
      description: Search for data points by pattern
      security:
        - bearerAuth: []
      parameters:
        - name: pattern
          in: query
          schema:
            type: string
          description: Search pattern (wildcards supported)
          example: "ExampleDP_*"
        - name: dpType
          in: query
          schema:
            type: string
          description: Filter by data point type
        - name: ignoreCase
          in: query
          schema:
            type: boolean
          description: Case-insensitive search
      responses:
        '200':
          description: List of data points
          content:
            application/json:
              schema:
                type: object
                properties:
                  datapoints:
                    type: array
                    items:
                      type: string
                    example: ["ExampleDP_Arg1.", "ExampleDP_Arg2."]
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Data Points]
      summary: Create data point
      description: Create a new data point (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dpeName, dpType]
              properties:
                dpeName:
                  type: string
                  example: "MyDataPoint."
                dpType:
                  type: string
                  example: "ExampleDP_Arg"
                systemId:
                  type: integer
                  example: 1
                dpId:
                  type: integer
                  example: 100
      responses:
        '200':
          description: Data point created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Data Points - Value Operations
...
  /restapi/v1/datapoints/{dpName}:
    delete:
      tags: [Data Points]
      summary: Delete data point
      description: Delete a data point (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: dpName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Data point deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /restapi/v1/datapoints/{dpeName}/type:
    get:
      tags: [Data Points]
      summary: Get element type
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Element type
          content:
            application/json:
              schema:
                type: object
                properties:
                  elementType:
                    $ref: '#/components/schemas/ElementType'
...
  /restapi/v1/datapoint-types:
    get:
      tags: [Data Point Types]
      summary: List data point types
      security:
        - bearerAuth: []
      parameters:
        - name: pattern
          in: query
          schema:
            type: string
        - name: systemId
          in: query
          schema:
            type: integer
        - name: includeEmpty
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of types
          content:
            application/json:
              schema:
                type: object
                properties:
                  dpTypes:
                    type: array
                    items:
                      type: string

    post:
      tags: [Data Point Types]
      summary: Create data point type
      description: Create a new data point type (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startNode]
              properties:
                startNode:
                  $ref: '#/components/schemas/DpTypeNode'
      responses:
        '200':
          description: Type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /restapi/v1/datapoint-types/{dpt}/structure:
    get:
      tags: [Data Point Types]
      summary: Get type structure
      security:
        - bearerAuth: []
      parameters:
        - name: dpt
          in: path
          required: true
          schema:
            type: string
        - name: includeSubTypes
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Type structure
          content:
            application/json:
              schema:
                type: object
                properties:
                  structure:
                    $ref: '#/components/schemas/DpTypeNode'
  /restapi/v1/datapoint-types/{dpt}:
    put:
      tags: [Data Point Types]
      summary: Change data point type
      description: Modify existing data point type (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: dpt
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startNode]
              properties:
                startNode:
                  $ref: '#/components/schemas/DpTypeNode'
      responses:
        '200':
          description: Type changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    delete:
      tags: [Data Point Types]
      summary: Delete data point type
      description: Delete a data point type (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: dpt
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Type deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /restapi/v1/datapoint-types/{dpt}/references:
    get:
      tags: [Data Point Types]
      summary: Get type references
      description: Get references to other DPTs in a DPT
      security:
        - bearerAuth: []
      parameters:
        - name: dpt
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Type references
          content:
            application/json:
              schema:
                type: object
                properties:
                  dptNames:
                    type: array
                    items:
                      type: string
                  dpePaths:
                    type: array
                    items:
                      type: string
  /restapi/v1/datapoint-types/{reference}/usages:
    get:
      tags: [Data Point Types]
      summary: Get type usages
      description: Get all DPTs that contain a specific DPT as reference
      security:
        - bearerAuth: []
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Type usages
          content:
            application/json:
              schema:
                type: object
                properties:
                  dptNames:
                    type: array
                    items:
                      type: string
                  dpePaths:
                    type: array
                    items:
                      type: string

  # Tags
...